+++ sdk/eventhub/azure-eventhub/azure/eventhub/__init__.py
@@ -2,7 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-from .pyamqp import constants
+from ._pyamqp import constants
 from ._common import EventData, EventDataBatch
 from ._version import VERSION
