+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/_connection.py
