+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/_platform.py
