+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/aio/_connection_async.py
