+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/aio/_link_async.py
