+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/aio/_sasl_async.py
