+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/aio/_transport_async.py
