+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/authentication.py
