+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/cbs.py
