+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/client.py
