+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/constants.py
