+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/endpoints.py
