+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/error.py
