+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/link.py
