+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/management_link.py
