+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/management_operation.py
