+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/message.py
