+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/mgmt_operation.py
