+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/outcomes.py
