+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/performatives.py
