+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/sasl.py
