+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/sender.py
