+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/session.py
