+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/types.py
