+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/utils.py
